Поправил бэкап view таблиц, поправил переход по якорю в форуме

+++ app/Classes/Metrika.php
@@ -152,7 +152,7 @@ class Metrika
             ];
         }
 
-        $hits = session('hits', 1);
+        $hits = (int) session('hits', 1);
 
         $hitsUpdate = [
             'allhits' => DB::raw('allhits + ' . $hits),
